requirements.txt
tensorflow>=2.12
numpy
matplotlib
seaborn
scikit-learn
Pillow
flask
flask-cors
pydot
graphviz
